feature : #6 단건 상품 조회 로직의 서비스 테스트 이전"
