+++ mini-commerce/src/main/java/me/hero/minicommerce/item/application/port/in/ShowOneItemUseCase.java
@@ -1,0 +1,5 @@
+package me.hero.minicommerce.item.application.port.in;
+
+public interface ShowOneItemUseCase {
+
+}
